  \name{RmixmodCombi-internal} 
        \alias{Combi} 
        \alias{barplotMixCluster} 
        \alias{histMixCluster} 
        \alias{mixmodEntPlot}
        \alias{pcws2_reg}
        \alias{pcws3_reg}
        \alias{xlog}
        \alias{MixmodCombiSol-class}
        \title{Internal RmixmodCombi Functions} 
        \description{ 
          Internal RmixmodCombi functions 
        } 
        \details{ 
          These are not intended to be called by the user. 
        } 
        \keyword{ internal } 
    